moved merge method from ODataExpandAssociation to CommandFormatter
